Remove Exceptional; improve entity handling
Removes the explicit-exception package. I had previously used this
because the Monad instance for Either was an orphan.  As of base-4.6,
the Monad instance for Either is now defined in the same module as for
Either itself, so this problem has gone away.  As a result I see no
need to continue using explicit-exception, and ceasing its use removes
a dependency.

Also, changes handling of entities slightly to give better error
messages.  Some bnaks deliver broken OFX files that will cause the
parse to fail when it encounters an ampersand; the error message now
gives a hint about this.  I had considered handling this broken
behavior in this library, but I now believe that a better way to
handle this is for the client code to filter or replace ampersands in
the source for banks that are known offenders in this area.  This
keeps the OFX parsing simpler and clean.